orderService.go and orderHandler.go

+++ internal/model/responseModels.go
@@ -45,6 +45,6 @@ type OrderItemResponse struct {
 	Id        int64   `json:"id"`
 	OrderId   int64   `json:"order_id"`
 	ProductId int64   `json:"product_id"`
-	Quantity  int64   `json:"quantity"`
+	Quantity  int     `json:"quantity"`
 	Price     float64 `json:"price"`
 }
